Clear cached provider properties after transition
A transition resource does the equivalent of an out-of-band
`puppet resource ...` operation. This means that the catalog will end up
containg a resource with a stale state if pre-fetching was triggered before
the transition. An example of this is a package resource in the catalot that
the agent thinks is present, because it was when pre-fetch happened, but is
actually absent because of a Transition. This patch clears cached state from
the catalog resource so that it will be re-fetched when the resource
is evaluated.